feat(landing): refactor DEFAULT_LANDING_MODULES into a separate definition file and update imports
