+++ src/modules/landing/definitions/defaultLandingModules.ts
@@ -1,0 +1,9 @@
+export const DEFAULT_LANDING_MODULES = [
+    { name: 'hero', order: 1, enabled: true },
+    { name: 'stats', order: 2, enabled: true },
+    { name: 'featuredCommands', order: 3, enabled: true },
+    { name: 'features', order: 4, enabled: true },
+    { name: 'partners', order: 5, enabled: true },
+    { name: 'faqs', order: 6, enabled: true },
+    { name: 'cta', order: 7, enabled: true },
+];
