+++ src/modules/landing/index.ts
@@ -1,16 +1,6 @@
 import { Module, ServiceContainer } from "zumito-framework";
 import { NavigationService } from "@zumito-team/admin-module/services/NavigationService.js";
 
-export const DEFAULT_LANDING_MODULES = [
-    { name: 'hero', order: 1, enabled: true },
-    { name: 'stats', order: 2, enabled: true },
-    { name: 'featuredCommands', order: 3, enabled: true },
-    { name: 'features', order: 4, enabled: true },
-    { name: 'partners', order: 5, enabled: true },
-    { name: 'faqs', order: 6, enabled: true },
-    { name: 'cta', order: 7, enabled: true },
-];
-
 export class LandingModule extends Module {
     requeriments = {
         services: ["NavigationService"],
@@ -35,8 +25,12 @@ export class LandingModule extends Module {
                     showDropdown: false,
                     sections: [
                         {
+                            id: 'landing',
                             label: 'Landing',
-                            items: [ { label: 'Modules', url: '/admin/landing/modules' } ],
+                            items: [ 
+                                { label: 'Modules', url: '/admin/landing/modules' },
+                                { label: 'Featured Commands', url: '/admin/landing/featured-commands' },
+                            ],
                         },
                     ],
                 },
